pyproject.toml
[project]
name = "onc-llm"
version = "0.1.0"
description = "reference implementation for oncologic history summarization."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Karan H. Jani" }]
dependencies = [
  "dspy-ai>=2.6.0",
  "pydantic>=2.10.0",
  "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
